refactor: Deleted some unnecessary comments, added a constant
